Feat/database (#2)
* Adding database integration when running locally

* Wiring up database and view

+++ src/main/java/com/course/insurancenotes/notes/Note.java
@@ -3,10 +3,10 @@ package com.course.insurancenotes.notes;
 
 public class Note{
 
-    private Integer userId;
+    private Long userId;
     private String note;
 
-    public Note(Integer userId, String note) {
+    public Note(Long userId, String note) {
         this.userId = userId;
         this.note = note;
     }
@@ -22,11 +22,11 @@ public class Note{
     public void setNote(String note) {
         this.note = note;
     }
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
